Replace remaining WinAPI types with standard C++ equivalents
DWORD→uint32_t, BYTE→uint8_t, HRESULT→int32_t, HANDLE→void*, UINT→uint32_t, INT→int32_t, WORD→uint16_t, LONG→int32_t, SHORT→int16_t, LONGLONG→int64_t, ULONG_PTR→uintptr_t, PBYTE→uint8_t*, LPWSTR/PWSTR→wchar_t*, FLOAT→float, CHAR→char, boolean→bool, CONST→const, TRUE→true, FALSE→false across 176 files (excluding vendor libs and Linux stubs).

+++ Minecraft.Client/Network/ServerChunkCache.cpp
@@ -166,9 +166,9 @@ LevelChunk* ServerChunkCache::create(
                 (int64_t*)&cache[idx], (int64_t)chunk, (int64_t)lastChunk) ==
             (int64_t)lastChunk)
 #else
-        if (InterlockedCompareExchangeRelease((LONG*)&cache[idx], (LONG)chunk,
-                                              (LONG)lastChunk) ==
-            (LONG)lastChunk)
+        if (InterlockedCompareExchangeRelease((int32_t*)&cache[idx], (int32_t)chunk,
+                                              (int32_t)lastChunk) ==
+            (int32_t)lastChunk)
 #endif
         {
             // Successfully updated the cache
